-Cast wadcount and patchcount to a BYTE when reading the WADs and DEHs used on a server to ensure values with the high bit set are not interpreted as negative numbers.
SVN r3672 (trunk)

+++ client/src/cl_main.cpp
@@ -3045,14 +3045,14 @@ void CL_LoadMap(void)
 	std::vector<std::string> newwadfiles, newwadhashes;
 	std::vector<std::string> newpatchfiles, newpatchhashes;
 
-	int wadcount = MSG_ReadByte();
+	int wadcount = (byte)MSG_ReadByte();
 	while (wadcount--)
 	{
 		newwadfiles.push_back(MSG_ReadString());
 		newwadhashes.push_back(MSG_ReadString());
 	}
 
-	int patchcount = MSG_ReadByte();
+	int patchcount = (byte)MSG_ReadByte();
 	while (patchcount--)
 	{
 		newpatchfiles.push_back(MSG_ReadString());
